Refactor: Decouple plugin/modules/http/client.py helpers
- Moved `format_error_message`, `_format_http_error_response`, `format_error_for_display`, and `is_audio_unsupported_error` to `plugin/modules/http/errors.py`.
- Moved `get_unverified_ssl_context`, `get_verified_ssl_context`, `_is_certificate_verify_error`, and `_is_local_host` to `plugin/modules/http/ssl_helpers.py`.
- Moved `iterate_sse`, `_extract_thinking_from_delta`, `_normalize_message_content`, and `_normalize_delta` to `plugin/modules/http/stream_normalizer.py`.
- Moved `sync_request` to `plugin/modules/http/requests.py`.
- Updated all import references across the entire codebase and test suite.
