+++ plugin/contrib/aihordeclient/__init__.py
@@ -31,7 +31,8 @@ from urllib.error import HTTPError, URLError
 from urllib.request import Request
 from plugin.framework.translation_tool import opustm_hf_translate, OPUSTM_SOURCE_LANGUAGES  # noqa F401
 
-from plugin.modules.http.client import sync_request, format_error_message
+from plugin.modules.http.requests import sync_request
+from plugin.modules.http.errors import format_error_message
 from plugin.framework.logging import log_exception
 from plugin.framework.constants import USER_AGENT
