+++ plugin/framework/translation_tool.py
@@ -25,7 +25,7 @@
 # https://github.com/ikks/aihorde-client/blob/main/LICENSE
 
 import json
-from plugin.modules.http.client import sync_request
+from plugin.modules.http.requests import sync_request
 
 API_TRANSLATE_GRADIO = "https://igortamara-opus-translate.hf.space/call/translate"
