+++ plugin/prompt_function.py
@@ -147,7 +147,7 @@ class PromptFunction(unohelper.Base, XPromptFunction):
                 from plugin.framework.async_stream import run_blocking_in_thread
                 return run_blocking_in_thread(self.ctx, self.client.chat_completion_sync, messages, max_tokens=max_tokens)
             except Exception as e:
-                from plugin.modules.http.client import format_error_for_display
+                from plugin.modules.http.errors import format_error_for_display
                 log.error("PROMPT error: %s" % str(e))
                 return format_error_for_display(e)
         return ""
